shop/requirements.txt
certifi==2019.9.11
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.7
django-allauth==0.40.0
django-filebrowser-no-grappelli==3.7.8
idna==2.8
oauthlib==3.1.0
Pillow==6.2.1
python-slugify==4.0.0
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
sqlparse==0.3.0
text-unidecode==1.3
urllib3==1.25.7
